refactor(common): extract buildClipboardData and insertClipboardData fix copy arrow-line error
