+++ .changeset/bright-otters-battle.md
@@ -1,0 +1,10 @@
+---
+'@plait/common': minor
+'@plait/draw': minor
+---
+
+extract buildClipboardData and insertClipboardData to common and support scalable fun
+
+rewrite buildClipboardData and insertClipboardData by basic method
+
+fix copy arrow-line error
