+++ packages/common/src/utils/index.ts
@@ -15,3 +15,4 @@ export * from './rotate';
 export * from './elements';
 export * from './animate';
 export * from './stroke';
+export * from './clipboard';
