+++ packages/draw/src/utils/clipboard.ts
@@ -1,16 +1,12 @@
-import { PlaitBoard, Point, Transforms, getElementById, idCreator } from '@plait/core';
-import { PlaitArrowLine, PlaitDrawElement, PlaitGeometry, PlaitShapeElement } from '../interfaces';
-import { PlaitImage } from '../interfaces/image';
+import { PlaitBoard, PlaitElement, Point, Transforms, getElementById, idCreator } from '@plait/core';
+import { buildClipboardData as basicBuildClipboard, insertClipboardData as basicInsertClipboard } from '@plait/common';
+import { PlaitArrowLine, PlaitDrawElement, PlaitGeometry } from '../interfaces';
 import { getConnectionPoint } from './arrow-line/arrow-line-common';
 import { PlaitTable } from '../interfaces/table';
 import { updateCellIds, updateRowOrColumnIds } from './table';
 
 export const buildClipboardData = (board: PlaitBoard, elements: PlaitDrawElement[], startPoint: Point) => {
-    return elements.map(element => {
-        if (PlaitDrawElement.isShapeElement(element)) {
-            const points = element.points.map(point => [point[0] - startPoint[0], point[1] - startPoint[1]]);
-            return { ...element, points } as PlaitGeometry;
-        }
+    return basicBuildClipboard(board, elements, startPoint, (element: PlaitElement) => {
         if (PlaitDrawElement.isArrowLine(element)) {
             let source = { ...element.source };
             let target = { ...element.target };
@@ -35,17 +31,12 @@ export const buildClipboardData = (board: PlaitBoard, elements: PlaitDrawElement
             points = points.map(point => [point[0] - startPoint[0], point[1] - startPoint[1]]);
             return { ...element, points, source, target } as PlaitArrowLine;
         }
-        return element;
+        return undefined;
     });
 };
 
 export const insertClipboardData = (board: PlaitBoard, elements: PlaitDrawElement[], startPoint: Point) => {
-    const idsMap: Record<string, string> = {};
-    elements.forEach(element => {
-        idsMap[element.id] = idCreator();
-    });
-    elements.forEach(element => {
-        element.id = idsMap[element.id];
+    basicInsertClipboard(board, elements, startPoint, (element: PlaitElement, idsMap: Record<string, string>) => {
         if (PlaitDrawElement.isArrowLine(element)) {
             if (element.source.boundId) {
                 const boundElement = elements.find(item => [element.source.boundId, idsMap[element.source.boundId!]].includes(item.id));
@@ -65,8 +56,5 @@ export const insertClipboardData = (board: PlaitBoard, elements: PlaitDrawElemen
             updateRowOrColumnIds(element as PlaitTable, 'column');
             updateCellIds(element.cells);
         }
-        element.points = element.points.map(point => [startPoint[0] + point[0], startPoint[1] + point[1]]) as [Point, Point];
-        Transforms.insertNode(board, element, [board.children.length]);
     });
-    Transforms.addSelectionWithTemporaryElements(board, elements);
 };
